+++ src/mchi112/hga/Point.java
@@ -2,25 +2,25 @@ package mchi112.hga;
 
 public class Point {
 
-    private int x;
-    private int y;
+    private float x;
+    private float y;
 
-    public Point(int x, int y) {
+    public Point(float x, float y) {
         this.x = x;
         this.y = y;
     }
 
-    public int X() {
+    public float X() {
         return x;
     }
 
-    public int Y() {
+    public float Y() {
         return y;
     }
 
     public float getDistance(Point p) {
-        int xDistance = Math.abs(X() - p.X());
-        int yDistance = Math.abs(Y() - p.Y());
+        float xDistance = Math.abs(X() - p.X());
+        float yDistance = Math.abs(Y() - p.Y());
         double distance = Math.sqrt( (xDistance*xDistance) + (yDistance*yDistance) );
 
         return (float)distance;
